Rename quad and tile art files to be consistent spelling
