+++ src/game/editor/quad_art.cpp
@@ -1,4 +1,4 @@
-#include "quadart.h"
+#include "quad_art.h"
 
 #include "editor.h"
 #include "editor_actions.h"
