+++ src/game/editor/quad_art.h
